requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
pymupdf
supabase
fpdf2==2.7.8
matplotlib
